Move around some exceptions to their appropriate locations
