+++ lib/Doctrine/ORM/Configuration.php
@@ -16,8 +16,8 @@ use Doctrine\ORM\Cache\Exception\MetadataCacheNotConfigured;
 use Doctrine\ORM\Cache\Exception\MetadataCacheUsesNonPersistentCache;
 use Doctrine\ORM\Cache\Exception\QueryCacheNotConfigured;
 use Doctrine\ORM\Cache\Exception\QueryCacheUsesNonPersistentCache;
-use Doctrine\ORM\Exception\InvalidEntityRepository;
-use Doctrine\ORM\Exception\ProxyClassesAlwaysRegenerating;
+use Doctrine\ORM\Configuration\Exception\InvalidEntityRepository;
+use Doctrine\ORM\Configuration\Exception\ProxyClassesAlwaysRegenerating;
 use Doctrine\ORM\Mapping\ClassMetadataFactory;
 use Doctrine\ORM\Mapping\DefaultEntityListenerResolver;
 use Doctrine\ORM\Mapping\Driver\AnnotationDriver;
