+++ lib/Doctrine/ORM/Configuration/Exception/ConfigurationException.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Doctrine\ORM\Exception;
+namespace Doctrine\ORM\Configuration\Exception;
+
+use Doctrine\ORM\Exception\ORMException;
 
 interface ConfigurationException extends ORMException
 {
