+++ lib/Doctrine/ORM/Configuration/Exception/InvalidEntityRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Doctrine\ORM\Exception;
+namespace Doctrine\ORM\Configuration\Exception;
 
 use Doctrine\Common\Persistence\ObjectRepository;
