+++ lib/Doctrine/ORM/Configuration/Exception/UnknownEntityNamespace.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Doctrine\ORM\Exception;
+namespace Doctrine\ORM\Configuration\Exception;
 
 use function sprintf;
