+++ lib/Doctrine/ORM/Persisters/Exception/InvalidOrientation.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Persisters\Exception;
 
-use Doctrine\ORM\Exception\PersisterException;
-
 final class InvalidOrientation extends \LogicException implements PersisterException
 {
     public static function fromClassNameAndField(string $className, string $field) : self
