+++ lib/Doctrine/ORM/Persisters/Exception/MatchingAssociationFieldRequiresObject.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Doctrine\ORM\Persisters;
+namespace Doctrine\ORM\Persisters\Exception;
 
-use Doctrine\ORM\Exception\PersisterException;
+use Doctrine\ORM\Persisters\Exception\PersisterException;
 use function sprintf;
 
 final class MatchingAssociationFieldRequiresObject extends \LogicException implements PersisterException
