+++ lib/Doctrine/ORM/Persisters/Exception/PersisterException.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Doctrine\ORM\Exception;
+namespace Doctrine\ORM\Persisters\Exception;
+
+use Doctrine\ORM\Exception\ORMException;
 
 interface PersisterException extends ORMException
 {
