+++ lib/Doctrine/ORM/Persisters/Exception/UnrecognizedField.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Persisters\Exception;
 
-use Doctrine\ORM\Exception\PersisterException;
 use function sprintf;
 
 final class UnrecognizedField extends \LogicException implements PersisterException
