+++ lib/Doctrine/ORM/Repository/Exception/RepositoryException.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Doctrine\ORM\Exception;
+namespace Doctrine\ORM\Repository\Exception;
+
+use Doctrine\ORM\Exception\ORMException;
 
 /**
  * This interface should be implemented by all exceptions in the Repository
