+++ lib/Doctrine/ORM/Tools/Exception/SchemaToolException.php
